requirements.txt
kafka-python==2.0.2
pandas
torch>=2.0.0
transformers>=4.30.0
peft>=0.4.0
pyspark==4.0.0
safetensors>=0.3.0
accelerate>=0.20.0
hf_xet
dotenv
langchain_google_genai
requests
apache-airflow